Cast cube viewport size to uint32 for gls.Scissor
gls.Scissor takes (int32 x, int32 y, uint32 w, uint32 h) — gls.Viewport
takes int32 throughout. Previously worked because cubeViewportSize was
an untyped int constant that auto-converted; the Retina-fix commit made
the size an explicit int32 (cubeSizePhys), which won't auto-convert
across the int32→uint32 boundary, so the build broke on macOS.

+++ gcode_viewer_v3/internal/ui/window.go
@@ -199,7 +199,7 @@ func Run(initialPath string) {
 		cubeX := int32(fbW) - cubeSizePhys
 		cubeY := int32(fbH) - cubeSizePhys - toolbarPhys
 		win.Gls().Enable(gls.SCISSOR_TEST)
-		win.Gls().Scissor(cubeX, cubeY, cubeSizePhys, cubeSizePhys)
+		win.Gls().Scissor(cubeX, cubeY, uint32(cubeSizePhys), uint32(cubeSizePhys))
 		win.Gls().Viewport(cubeX, cubeY, cubeSizePhys, cubeSizePhys)
 		win.Gls().Clear(gls.DEPTH_BUFFER_BIT)
 		if err := rend.Render(state.cube.Scene, state.cube.Camera); err != nil {
